subreddit_stream/requirements.txt
praw==7.7.0
protobuf==4.22.1
python-dotenv==1.0.0
google-api-python-client==2.84.0
google-cloud-secret-manager==2.16.1
google-cloud-storage==2.8.0